advisor does not belong to network
