+++ internal/components/azaks/registry_test.go
@@ -6,11 +6,11 @@ import (
 
 func TestGetReadOnlyAzCommands_ContainsNodepoolCommands(t *testing.T) {
 	commands := GetReadOnlyAzCommands()
-	
+
 	// Check that nodepool list command is included
 	foundNodepoolList := false
 	foundNodepoolShow := false
-	
+
 	for _, cmd := range commands {
 		if cmd.Name == "az aks nodepool list" {
 			foundNodepoolList = true
@@ -31,11 +31,11 @@ func TestGetReadOnlyAzCommands_ContainsNodepoolCommands(t *testing.T) {
 			}
 		}
 	}
-	
+
 	if !foundNodepoolList {
 		t.Error("Expected to find 'az aks nodepool list' command in read-only commands")
 	}
-	
+
 	if !foundNodepoolShow {
 		t.Error("Expected to find 'az aks nodepool show' command in read-only commands")
 	}
@@ -48,30 +48,30 @@ func TestRegisterAzCommand_NodepoolCommands(t *testing.T) {
 		Description: "List node pools in a managed Kubernetes cluster",
 		ArgsExample: "--cluster-name myAKSCluster --resource-group myResourceGroup",
 	}
-	
+
 	tool := RegisterAzCommand(listCmd)
-	
+
 	if tool.Name != "az_aks_nodepool_list" {
 		t.Errorf("Expected tool name 'az_aks_nodepool_list', got '%s'", tool.Name)
 	}
-	
+
 	if tool.Description == "" {
 		t.Error("Expected tool description to be set")
 	}
-	
+
 	// Test that nodepool show command can be registered
 	showCmd := AksCommand{
 		Name:        "az aks nodepool show",
 		Description: "Show the details for a node pool in the managed Kubernetes cluster",
 		ArgsExample: "--cluster-name myAKSCluster --resource-group myResourceGroup --name nodepool1",
 	}
-	
+
 	tool2 := RegisterAzCommand(showCmd)
-	
+
 	if tool2.Name != "az_aks_nodepool_show" {
 		t.Errorf("Expected tool name 'az_aks_nodepool_show', got '%s'", tool2.Name)
 	}
-	
+
 	if tool2.Description == "" {
 		t.Error("Expected tool description to be set")
 	}
